api/main_service/serializers/city_serializers.py
from rest_framework import serializers
from main_service.models import City


class CitySerializer(serializers.ModelSerializer):
    class Meta:
        model = City
        fields = ("id", "name", "country")